kycagents/pyproject.toml
[project]
name = "kycagents"
version = "0.1.0"
description = "KycAgents using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.9.3",
    "python-dotenv>=1.0.1",
    "litellm>=1.41.2",
    "fastapi>=0.128.5",
    "pdf2image>=1.17.0",
    "pillow>=10.3.0",
    "requests>=2.31.0",
]

[project.scripts]
kycagents = "kycagents.main:run"
run_crew = "kycagents.main:run"
train = "kycagents.main:train"
replay = "kycagents.main:replay"
test = "kycagents.main:test"
run_with_trigger = "kycagents.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
